Fix link failure when building on Fedora 64bit
Fedora and other Linux distributions tend to use `lib64` as the library
directory and when building and installing the native library with CMake the
directory structure in the prefix conforms to this standard.

This makes the linking fail since the search directory for `libnng` is set to
`lib`.

This patch adds `lib64` to the library search path on Linux.

The same issue is also present for `mbedtls` and was fixed there as well.